multiAV/requirements.txt
Django>=3.2.0
requests
python-dotenv
load_dotenv